Add output for error handling (#2)
* Add output for error handling

This adds output for errors, previously it was `fmt.Sprintf` which will only
format and doesn't output the string.

The commit also provides more information about when the command isn't run in
the git toplevel. This could be solved by using `git rev-parse
--show-toplevel` but as far as I can tell, this isn't supported in go-git
right now.

* Remove extra checking for .git directory

This removes a check that for the .git directory which is also done in go-git.